rename Slider to BaseSlider
