+++ components/atoms/BaseSlider.vue
